% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc.nonwear.R
\name{acc.nonwear}
\alias{acc.nonwear}
\title{Identification of accelerometer non-wear time}
\usage{
acc.nonwear(dataset, epoch.length, window.length, interruption.length)
}
\arguments{
\item{dataset}{The merged accelerometer and gps data}

\item{epoch.length}{The length of epoch used, in seconds}

\item{window.length}{The length of window you want non-wear time to be measured , in minutes}

\item{interruption.length}{The lenth of any window within the non-wear time that can contain counts, in minutes}
}
\value{
A data.frame of the merged data that was input,
 but with non wear time set as NA
}
\description{
Returns a dataset in the same format as the input but with accelerometer non-wear
time marked as NA
}
\details{
The function assumes you have used the \code{\link{gps.acc.merger}} function.
If you have not used that function, your accelerometer variables need to be
labelled "Axis1", "Axis2" and "Axis3".
It should work for 1, 2 or 3 accelerometer axes, it will use as many as are present
}
