% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.matrix.R
\name{confusion.matrix}
\alias{confusion.matrix}
\title{Confusion matrix}
\usage{
confusion.matrix(predicted, observed)
}
\arguments{
\item{predicted}{A factor variable of predicted mode}

\item{A}{factor variable of observed mode}
}
\value{
A matrix of hwo predicted and observed modes correspond
}
\description{
Returns a confusion matrix, based on a predicted and observed variable
}
\details{
The output rows are the predicted modes, the output columns are the observed
}
\examples{
pred<-numeric(10)+1
pred[5:6]<-2
pred<-factor(pred,labels=c("Mode1","Mode2"))

obs<-numeric(10)+1
obs[6:8]<-2
obs<-factor(obs,labels=c("Mode1","Mode2"))

confusion.matrix(pred,obs)
}
