% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.loss.gps.R
\name{data.loss.gps}
\alias{data.loss.gps}
\title{Data loss due to GPS cleaning}
\usage{
data.loss.gps(speed.cutoff, hdop.cutoff, neighbour.number, neighbour.window,
  epoch.length, dataset)
}
\arguments{
\item{speed.cutoff}{The cut-off above which speed is deemed unreliable, in the same units as your speed variable}

\item{hdop.cutoff}{The cut-off above which your horizontal dilution of precision is deemed unreliable}

\item{neighbour.number}{an integer: if you are removing isolated points, how few neighbours counts as isolated}

\item{neighbour.window}{The width of a window in which to look for neighbours to define whether a point is
isolated, in seconds}

\item{dataset}{The dataset you need cleaned}
}
\description{
Returns a two column data.frame, with the data amount lost at each level
of data cleaning, in epochs
}
\details{
Set up exactly the same as \code{\link{gps.cleaner}}, this function
tell you how much data you lose from cleaning the data with your given rules
}
