% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.busmod.R
\docType{data}
\name{fitted.busmod}
\alias{fitted.busmod}
\title{The fitted model from Procter et al 2018 including buses}
\format{A xgb.Booster, fitted xgboost model}
\usage{
fitted.busmod
}
\description{
A fitted xgboost model with 6 travel modes (walk, cycle, vehicle, train, bus, stationary)
, to allow prediction to other datasets
}
\keyword{datasets}
