% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.fulmod.R
\docType{data}
\name{fitted.fullmod}
\alias{fitted.fullmod}
\title{The full fitted model from Procter et al 2018}
\format{A xgb.Booster, fitted xgboost model}
\usage{
fitted.fullmod
}
\description{
A fitted xgboost model with 5 travel modes (walk, cycle, vehicle, train, stationary)
, to allow prediction to other datasets
}
\keyword{datasets}
