% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foldpred.R
\name{foldpred}
\alias{foldpred}
\title{Predict modal travel mode from cross-validated model}
\usage{
foldpred(data, model)
}
\arguments{
\item{data}{The data to predict to, make sure the correct variables are selected using \code{\link{pred.data}}}

\item{model}{The cross-validated model, created with \code{\link{cross.validator}}}
}
\value{
A factor variable of the same length as variables in the data, with predicted travel mode
}
\description{
Returns a factor of the length of the data with predicted travel mode
}
