% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gps.cleaner.R
\name{gps.cleaner}
\alias{gps.cleaner}
\title{Cleaning of GPS data}
\usage{
gps.cleaner(speed.cutoff, hdop.cutoff, neighbour.number, neighbour.window,
  epoch.length, dataset)
}
\arguments{
\item{speed.cutoff}{The cut-off above which speed is deemed unreliable, in the same units as your speed variable}

\item{hdop.cutoff}{THe cut-off above which your horizontal dilution of precision is deemed unreliable}

\item{neighbour.number}{an integer: if you are removing isolated points, how few neighbours counts as isolated}

\item{neighbour.window}{The width of a window in which to look for neighbours to define whether a point is
isolated, in seconds}

\item{dataset}{The dataset you need cleaned}
}
\value{
A data.frame in the same format as your input, but with GPS related variables set as NA
}
\description{
Returns a data.frame in the same format as the input
but with points deemed unreliable set as NA
}
\details{
Cleans unreliable points base on speed, HDOP and number of neighbours.
Assumes you have used the \code{\link{gps.acc.merger}} function, and your data is set
up in that manner. If you have not that is not a problem, but the function expects
variables named "longitude","latitude","speed","hdop","pdop","vdop" and "sumsnr".
If you want all of those things set as NA, you need the variables labelled correctly
}
