% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/near.train.R
\name{near.train}
\alias{near.train}
\title{Distance from all points to the nearest train line}
\usage{
near.train(dataset, trainline.psp, trainline.p4s)
}
\arguments{
\item{dataset}{a merged gps/accelerometer dataset}

\item{trainline.psp}{a \code{psp} (line segement pattern from the package \link[package]{spatstat})
of train lines}

\item{trainline.p4s}{the \code{proj4string} of the trainline}
}
\description{
Measures euclidean distance from each point to the nearest trainline
}
\details{
We assume you have processed accelerometer data using the \code{\link{process.acc}}
and \code{\link{gps.acc.merge}}. This function then takes train line data, in \code{psp} format, 
and uses the \code{nncross} function from the \link[package]{spatstat} package to measure distance from 
each point to the nearest train line.

To find out the \code{proj4string} of the trainline data, use package \link[package]{sp} function 
\code{proj4string} on a \code{SpatialLinesDataFrame} of train data
}
