% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw.predict.R
\name{nw.predict}
\alias{nw.predict}
\title{non-wear predictor}
\usage{
nw.predict(dataset, right = TRUE)
}
\arguments{
\item{dataset}{The data.frame you want nonwear predicted for}
}
\value{
Returns a variable, the same length as those in the input dataset, which is 0 if wear time
and 1 if nonwear
}
\description{
Uses a supervised machine learning model to predict non wear time for hip worn accelerometers
}
