% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwmod.R
\docType{data}
\name{nwmod}
\alias{nwmod}
\title{The fitted model to predict nonwear}
\format{A xgb.Booster, fitted xgboost model}
\usage{
nwmod
}
\description{
A fitted xgboost model with 2 categories (0 = wear time, 1=nonwear)
}
\keyword{datasets}
