% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.summary.R
\name{output.summary}
\alias{output.summary}
\title{Create summaries of processed files}
\usage{
output.summary(folder_location)
}
\arguments{
\item{folder_location}{the location fo the folders, the rest of the inputs have to be edited in 
the output_options.csv in that folder}
}
\description{
summarises a folder worth of processed files by day and week
}
\details{

}
