% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overall.acc.R
\name{overall.acc}
\alias{overall.acc}
\title{Overall accuracy of crossvalidated model}
\usage{
overall.acc(cvd.model, full.dataset)
}
\arguments{
\item{cvd.model}{a cross-validated model, created with \code{\link{cross.validator}}}

\item{full.dataset}{the full dataset to predict to}
}
\description{
The overall confusion matrix and model accuracy scores
}
\details{
Takes a cross-validated model, created with \code{\link{cross.validator}}
and outputs the overall confusion matix and model accuracy scores
}
