% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post2.R
\name{post2}
\alias{post2}
\title{Post processing: local patterns}
\usage{
post2(pred.variable, dataset)
}
\arguments{
\item{pred.variable}{The variable you want processed, a factor}

\item{dataset}{the dataset this is based on}
}
\value{
A variable, of the same format as \code{pred.variable}, with all points made the same
as those which have over 50% of points assigned to them within 4 minutes.
}
\description{
Returns a variable, which identifies over 50% of neighbours within 4 minutes
they are of the same mode, and switches the central point to that mode
}
\details{
We assume that if one mode is dominant both before and after the central point,
then that mode is likely what the central point should be. We therefore take a count of
each mode from two minutes before this point to two minutes after this point. If
a single mode has over half the points within the window, then the central point it
assigned to that mode.
}
