% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.data.R
\name{pred.data}
\alias{pred.data}
\title{selection of veriables neccessary for prediction}
\usage{
pred.data(merged.data)
}
\arguments{
\item{merged.data}{a dataset of merged data with rolling averages calculated}
}
\description{
Takes merged gps/acc data, with rolling averages calculated, subsets to only 
those variables that are needed to determine travel mode
}
\details{
Here we assume you have processed acceleormeter data using \code{\link{process.acc}},
 \code{\link{gps.acc.merge}} and \code{\link{rollav.calc}}.
 
 This funciton then selects only the necessary variables for prediction, and outputs them as a matrix, 
 ready to be fed to a fitted xgboost model
 
 @export
}
