% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.folder.R
\name{process.folder}
\alias{process.folder}
\title{Process a folder containing accelerometer and GPS data}
\usage{
process.folder(folder_location)
}
\arguments{
\item{folder_location}{the location fo the folders, the rest of the inputs have to ve edited in the input_options.csv in that folder}
}
\description{
writes a folder worth of .csv files of processed data
}
\details{
Currently only tested with Qstarz GPS device files. If you need other types contect the author
, they can be included with ease
}
