% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rollav.calc.R
\name{rollav.calc}
\alias{rollav.calc}
\title{Calculate rolling averages from a dataset}
\usage{
rollav.calc(dataset, acc = TRUE, gps = TRUE)
}
\arguments{
\item{dataset}{A merged gps/acc dataset}

\item{acc}{whether there is accelerometer data, default = TRUE}

\item{gps}{whether there is GPS data, default=TRUE}
}
\description{
Takes a merged gps/acc dataset and returns the same dataset with added rolling average variables
}
\details{
Here we assume you have processed accelerometer data using \code{\link{process.acc}}, 
merged it to GPS data using \code{\link{gps.acc.merge}}, and added variables called \emph{near.train}, 
\emph{dist.next.min} and \emph{dist.last.min}, possibly using \code{\link{near.train}} and 
\code{\link{distance.moved}}

This function will then take that dataset, and output the same dataset with
 the rolling averages neccessary to fit the model in Procter et al. 2018 added to it
}
