% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ug.journeys.R
\name{ug.journeys}
\alias{ug.journeys}
\title{Underground journey identifier}
\usage{
ug.journeys(dataset, station.ppp)
}
\arguments{
\item{dataset}{the dataset to check for missed tube journeys}

\item{station.ppp}{a dataset of underground stations to measure distance to, in the spatstat format .ppp}
}
\description{
Checks whether gaps in data start and end close to tube stations. If they do
and the gap lasts less than two hours, the missing data is marked as a missed tube journey
}
\details{
Firstly looks for gaps in data over 120 seconds and under 2 hours, then, for each start and end point 
of those gaps, it measures distance to the nearest underground station. If both start and end point are within
200m of an underground station, and it is not the same underground station, the missing data is marked as
a potential underground journey, with  a length of the gap in the data, in seconds
}
