% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.validator.R
\name{cross.validator}
\alias{cross.validator}
\title{Model cross-validation}
\usage{
cross.validator(training.dataset, formula, cv.marker)
}
\arguments{
\item{training.dataset}{The data.frame of training data you want to cross-validate}

\item{formula}{The formula of the model you want to use to fit the randomForest,
e.g. \code{true.mode~ax1.mean+spd.mean} would have the true mode being
classified by the mean of accelerometer axis 1 and mean speed}

\item{cv.marker}{A numeric marker for cross-validation subsets}
}
\value{
A four column matrix (I realise this is not terribly elegant, it will probably be changed).
The rows contain different cross-validation subsets.

The columns are as follows:

1. Fitted randomForest models for each cross-validation subset

2. Test data.frame for each cross-validation subset

3. The \code{\link{confusion.matrix}} for each cross-validation subset

4. The \code{\link{model.acc}} for each cross-validation subset
}
\description{
Returns a four column matrix, with as many rows as there are
crossvalidation subsets.
Column 1 has a fitted randomForest for each cross-valitation subset.
Column 2 has the test data.frame for each cross-validation subset.
Column 3 has the \code{\link{confusion.matrix}} for each cross-validation subset
Column 4 has the \code{\link{model.acc}} for each cross-validation subset
}
