% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.moved.R
\name{distance.moved}
\alias{distance.moved}
\title{Distance moved}
\usage{
distance.moved(dataset, last, time.window, epoch.length)
}
\description{
a function that measures distance moved between points separated by a time window
}
