% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.acc.R
\name{model.acc}
\alias{model.acc}
\title{Model accuracy}
\usage{
model.acc(conf.mat)
}
\arguments{
\item{conf.mat}{a confusion matrix generated by \code{\link{confusion.matrix}}}
}
\value{
A three column data.frame, with the precision and recall per mode
}
\description{
Returns model accuracy from a confusion matrix derived from
\code{\link{confusion.matrix}}
}
\details{
Return precision and recall for each mode.

Precision: the percentage of points which the model predicts as each mode
which are also observed as that mode

Recall: the percentage of points which are observed as each mode that the model
correctly predicts

 @examples
pred<-numeric(10)+1
pred[5:6]<-2
pred<-factor(pred,labels=c("Mode1","Mode2"))

obs<-numeric(10)+1
obs[6:8]<-2
obs<-factor(obs,labels=c("Mode1","Mode2"))

model.acc(confusion.matrix(pred,obs))
}
