% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post2.R
\name{post2}
\alias{post2}
\title{Post processing: local patterns}
\usage{
post2(pred.variable, epoch.length, window.width, prop.agreement)
}
\arguments{
\item{pred.variable}{The variable you want processed, either factor or numeric}

\item{epoch.length}{The length of the epoch you are using, in seconds}

\item{window.width}{How wide a window you want to smooth, in seconds, centered on a point}

\item{prop.agreement}{What proportion of that window you want to have the same mode}
}
\value{
A variable, of the same format as \code{pred.variable}, with all points made the same
as those which have over the specified proportion of the specified window.
}
\description{
Returns a variable, which identifies the majority mode within a specified
window and sets the center point as that mode
}
\details{
We assume that if one mode is dominant both before and after the central point,
then that mode is likely what the central point should be. We therefore take a count of
each mode from two minutes before this point to two minutes after this point. If
a single mode has over half the points within the window, then the central point it
assigned to that mode.
}
\examples{
eg.sequence<-rep(seq(1,3,1),each=6)
eg.sequence[8:10]<-1
eg.sequence<-factor(eg.sequence,labels=c("walk","run","cycle"))
eg.sequence
post2(pred.variable = eg.sequence, epoch.length = 10, window.width = 60, prop.agreement = 0.75)
}
