% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post3.R
\name{post3}
\alias{post3}
\title{Post processing: segmentation}
\usage{
post3(pred.variable, dataset, window.width, segment.length, epoch.length)
}
\arguments{
\item{pred.variable}{The variable containing predictions that you want to segment}

\item{dataset}{the dataset your prediction is contained within}

\item{window.width}{The time gap between points that dentes a new segment, in seconds}

\item{segment.length}{The minimum length of a segment in seconds}

\item{epoch.length}{The length of the epoch, in seconds}
}
\value{
A variable of post processed travel mode, in the same format as as
\code{pred.variable}
}
\description{
post3 identifies segements in the data, then assessed if one predicted mode
dominates this segement. If it does, post 3 sets the entire segement as
that mode.
}
